requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyjwt==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.47.2
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
